marketing-dashboard/requirements.txt
streamlit>=1.35.0
google-ads>=23.1.0
google-analytics-data>=0.18.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
facebook-business>=19.0.0
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0
plotly>=5.20.0
python-dotenv>=1.0.0
cryptography>=42.0.0
python-dateutil>=2.9.0
